fix: correct lint errors
