+++ frontend/src/context/AuthContext.tsx
@@ -1,15 +1,7 @@
-import { createContext, useContext, useState } from 'react';
+import { useState } from 'react';
 import type { ReactNode } from 'react';
+import { AuthContext } from './AuthContextValue';
 import type { AuthenticatedUser } from '../lib/api';
-
-interface AuthContextType {
-  user: AuthenticatedUser | null;
-  accessToken: string | null;
-  setSession: (accessToken: string, user: AuthenticatedUser) => void;
-  clearSession: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
 export function AuthProvider({ children }: { children: ReactNode }) {
   const [user, setUser] = useState<AuthenticatedUser | null>(null);
@@ -30,12 +22,4 @@ export function AuthProvider({ children }: { children: ReactNode }) {
       {children}
     </AuthContext.Provider>
   );
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 }
